Update to 1.17

+++ src/main/java/me/dustin/jex/helper/math/ClientMathHelper.java
@@ -132,19 +132,19 @@ public enum ClientMathHelper {
         double d0 = vec.x - vec1.x;
         double d1 = vec.y - vec1.y;
         double d2 = vec.z - vec1.z;
-        return MathHelper.sqrt(d0 * d0 + d1 * d1 + d2 * d2);
+        return MathHelper.sqrt((float)(d0 * d0 + d1 * d1 + d2 * d2));
     }
 
     public float getDistance2D(Vec2f vec, Vec2f vec1) {
         double d0 = vec.x - vec1.x;
         double d1 = vec.y - vec1.y;
-        return MathHelper.sqrt(d0 * d0 + d1 * d1);
+        return MathHelper.sqrt((float)(d0 * d0 + d1 * d1));
     }
 
     public float getDistance2D(Vec3d vec, Vec3d vec1) {
         double d0 = vec.x - vec1.x;
         double d1 = vec.z - vec1.z;
-        return MathHelper.sqrt(d0 * d0 + d1 * d1);
+        return MathHelper.sqrt((float)(d0 * d0 + d1 * d1));
     }
 
     public Vec3d getVec(Entity entity) {
